Submission/Root/Code/requirements.txt
alembic==1.13.2
amqp==5.2.0
APScheduler==3.10.4
bcrypt==4.0.1
billiard==4.2.0
blinker==1.8.2
cachelib==0.9.0
celery==5.4.0
chardet==5.2.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
DateTime==5.5
distlib==0.3.8
dnspython==2.6.1
email_validator==2.2.0
filelock==3.15.4
Flask==3.0.3
Flask-Caching==2.3.0
Flask-Cors==4.0.1
Flask-Excel==0.0.7
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-Principal==0.4.0
Flask-Security-Too==5.4.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.0.3
idna==3.7
importlib_resources==6.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
kombu==5.3.7
lml==0.1.0
Mako==1.3.5
MarkupSafe==2.1.5
packaging==24.1
passlib==1.7.4
platformdirs==4.2.2
plotly==5.23.0
prompt_toolkit==3.0.47
pyexcel==0.7.0
pyexcel-io==0.6.6
pyexcel-webio==0.1.4
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
redis==5.0.8
setuptools==71.1.0
six==1.16.0
SQLAlchemy==2.0.31
tenacity==9.0.0
texttable==1.7.0
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
vine==5.1.0
virtualenv==20.26.3
wcwidth==0.2.13
Werkzeug==3.0.3
WTForms==3.1.2
zope.interface==6.4.post2
